requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
plotly>=5.18
gspread>=6.0
google-auth>=2.28
requests>=2.31
openpyxl>=3.1
